onnx: fix references to tmdir in build paths
Recipes that use the ONNXConfig cmake will try to find protobuf headers
in the tmpdir used for this build if it's specified here. That won't
work if the workspace is cleaned (deleting tmp or rm_work).

Turns out they're not needed as the protobuf search paths will be
in the default locations, and cmake can provide its own prefix path,
so just remove these lines.

Signed-off-by: Kurt Kiefer <kekiefer@gmail.com>